--- packages/boxel-ui/types/ember-composable-helpers/helpers/compact.d.ts
@@ -1,9 +1,0 @@
-/* eslint-disable @typescript-eslint/ban-types */
-import Helper from '@ember/component/helper';
-
-export default class CompactHelper<T> extends Helper<{
-  Args: {
-    Positional: [Array<T> | T];
-  };
-  Return: Array<T>;
-}> {}
