--- packages/boxel-ui/types/ember-composable-helpers/helpers/noop.d.ts
@@ -1,8 +1,0 @@
-import Helper from '@ember/component/helper';
-
-interface NoopHelperSignature {
-  Args: { Positional: string[] };
-  Return: () => void;
-}
-
-export default class NoopHelper extends Helper<NoopHelperSignature> {}
