--- packages/boxel-ui/types/ember-composable-helpers/helpers/optional.d.ts
@@ -1,9 +1,0 @@
-/* eslint-disable @typescript-eslint/ban-types */
-import Helper from '@ember/component/helper';
-
-export default class OptionalHelper<T extends Function> extends Helper<{
-  Args: {
-    Positional: [func?: T];
-  };
-  Return: T;
-}> {}
