+++ packages/boxel-ui/types/ember-composable-helpers/index.d.ts
@@ -1,0 +1,11 @@
+declare module 'ember-composable-helpers/helpers/compact' {
+  /* eslint-disable @typescript-eslint/ban-types */
+  import Helper from '@ember/component/helper';
+
+  export default class CompactHelper<T> extends Helper<{
+    Args: {
+      Positional: [Array<T> | T];
+    };
+    Return: Array<T>;
+  }> {}
+}
